Keep API outside ErrorBoundary
